# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BuzJet API Documentation",
        "description": "Swagger documentation for BuzJet API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/packages": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get list of packages",
                "description": "Retrieve all packages along with their related data",
                "operationId": "6f73d13161427aaca37ac857421f9192",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Package"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Create a new package",
                "description": "Add a new package along with destinations and hotels",
                "operationId": "78cee803dbb936eb08f8ffc7e5c81a81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Package created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Package"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packages/{id}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get package by ID",
                "description": "Retrieve a specific package along with its related data",
                "operationId": "aaf4dd9ee9ffe63306c1e788d6fef3d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the package",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Package"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Update package by ID",
                "description": "Update the details of an existing package",
                "operationId": "05ca6c0a3eb75d7c09b8334718dbbc58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the package to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Package updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Package"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Delete package by ID",
                "description": "Remove a package and its relationships from the database",
                "operationId": "026af3ae89f479c04a753d9ff635748a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the package to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package deleted successfully"
                    },
                    "404": {
                        "description": "Package not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Package": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "duration": {
                        "type": "integer"
                    },
                    "night": {
                        "type": "integer"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "created_by": {
                        "type": "integer"
                    },
                    "destinations": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "hotels": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "user": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PackageInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "duration": {
                        "type": "integer"
                    },
                    "night": {
                        "type": "integer"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "created_by": {
                        "type": "integer"
                    },
                    "destination_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "hotel_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}